Renamed OpportunityCardGrid"
