+++ apps/charterafrica/src/lib/data/common/index.js
@@ -225,6 +225,7 @@ export async function processPageFellowships({ blocks }) {
       registerLink: {
         href: `/register/events/${i}`,
       },
+      registerText: "Register ",
       status: ["upcoming", "past"][Math.floor(Math.random() * 2)],
       featured: i === 0,
     })),
